requirements.txt
pyzmq>=26.0.0
pyttsx3>=2.90
